+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - no longer fetch dynamic options on read only forms
 
+### Fixed
+
+- submission PDF fileame in post submission receipt modal
+
 ## [8.10.0] - 2025-09-01
 
 ### Added